Moving radial profile class to more appropriate place
