+++ stellarphot/photometry/__init__.py
@@ -1,4 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
 from .photometry import *
+from .profiles import *
 from .source_detection import *
